added react-common
