+++ package.json
@@ -45,14 +45,15 @@
   "devDependencies": {
     "@types/graphql": "^0.9.0",
     "@types/history": "^3.2.0",
-    "@types/lodash": "^4.14.43",
     "@types/jest": "^19.2.2",
     "@types/jquery": "^2.0.41",
     "@types/jsdom": "^2.0.29",
+    "@types/lodash": "^4.14.43",
     "@types/material-ui": "^0.16.0",
     "@types/node": "^7.0.12",
-    "@types/react": "^15.0.20",
+    "@types/react": "^15.0.22",
     "@types/react-dom": "^0.14.23",
+    "@types/react-intl": "^2.2.6",
     "@types/react-redux": "^4.4.38",
     "@types/react-router": "^3.0.9",
     "@types/react-router-redux": "^4.0.43",
