+++ packages/react-common/package.json
@@ -59,7 +59,7 @@
     "@types/jest": "^19.2.2",
     "@types/jsdom": "^2.0.29",
     "@types/node": "^7.0.12",
-    "@types/react": "^15.0.20",
+    "@types/react": "^15.0.22",
     "@types/react-dom": "^0.14.23",
     "@types/react-redux": "^4.4.38",
     "apollo-client": "^1.0.0-rc.6",
