+++ packages/react-common/src/components/Match.tsx
@@ -1,0 +1,40 @@
+// import { State } from '../../common/types';
+// import * as React from 'react';
+// import { Match as ReactRouterMatch, Redirect } from 'react-router';
+// import { connect } from 'react-redux';
+
+// const haveAccess = (viewer, authorized) => authorized ? viewer : true;
+
+// const Match = (
+//   {
+//     authorized,
+//     component: Component,
+//     render,
+//     viewer,
+//     ...props
+//   },
+// ) => (
+//   <ReactRouterMatch
+//     {...props}
+//     render={renderProps =>
+//       haveAccess(viewer, authorized)
+//         ? render ? render(renderProps) : <Component {...renderProps} />
+//         : <Redirect
+//             to={{
+//               pathname: '/signin',
+//               state: { from: renderProps.location },
+//             }}
+//           />}
+//   />
+// );
+
+// Match.propTypes = {
+//   authorized: React.PropTypes.bool,
+//   component: React.PropTypes.any,
+//   render: React.PropTypes.func,
+//   viewer: React.PropTypes.object,
+// };
+
+// export default connect((state: State) => ({
+//   viewer: state.users.viewer,
+// }))(Match);
