+++ packages/react-common/src/components/index.tsx
@@ -3,8 +3,8 @@ export { default as Button } from './Button';
 export { default as Field } from './Field';
 export { default as Heading } from './Heading';
 export { default as Image } from './Image';
-export { default as Loading } from './Loading';
-export { default as Match } from './Match';
+// export { default as Loading } from './Loading';
+// export { default as Match } from './Match';
 export { default as Message } from './Message';
 export { default as OutlineButton } from './OutlineButton';
 export { default as PageHeader } from './PageHeader';
