+++ packages/redux-bootstrap/package.json
@@ -69,7 +69,7 @@
     "@types/jsdom": "^2.0.29",
     "@types/material-ui": "^0.16.0",
     "@types/node": "^7.0.12",
-    "@types/react": "^15.0.20",
+    "@types/react": "^15.0.22",
     "@types/react-dom": "^0.14.23",
     "@types/react-redux": "^4.4.38",
     "@types/react-router": "^3.0.9",
