+++ packages/redux-core/src/index.ts
@@ -5,3 +5,5 @@ export { buttonMessages } from './app/buttonMessages';
 export { linkMessages } from './app/linksMessages';
 export { isClient } from './app/isClient';
 export { isReactNative } from './app/isReactNative';
+export { appError, setTheme, appOnline, appShowMenu, toggleBaseline } from './app/actions';
+export { setCurrentLocale } from './intl/actions';
